+++ domain/src/main/java/com/remindme/domain/repository/TaskRepository.kt
@@ -47,5 +47,5 @@ interface TaskRepository {
      *
      * @return selected task
      */
-    suspend fun findTaskById(taskId: Int?): Task?
+    suspend fun findTaskById(taskId: Long?): Task?
 }
