+++ domain/src/main/java/com/remindme/domain/usecase/alarm/implementation/UpdateTaskAsRepeatingImpl.kt
@@ -10,7 +10,7 @@ class UpdateTaskAsRepeatingImpl @Inject constructor(
     private val taskRepository: TaskRepository
 ) : UpdateTaskAsRepeating {
 
-    override suspend operator fun invoke(taskId: Int?, interval: AlarmInterval?) {
+    override suspend operator fun invoke(taskId: Long?, interval: AlarmInterval?) {
         val task = taskRepository.findTaskById(taskId) ?: return
         logger.debug("UpdateTaskAsRepeating = Task = '${task.title} as '$interval")
