+++ domain/src/main/java/com/remindme/domain/usecase/task/LoadTask.kt
@@ -13,5 +13,5 @@ interface LoadTask {
      *
      * @return observable to be subscribe
      */
-    suspend operator fun invoke(taskId: Int?): Task?
+    suspend operator fun invoke(taskId: Long?): Task?
 }
