+++ domain/src/main/java/com/remindme/domain/usecase/task/implementation/UpdateTaskTitleImpl.kt
@@ -13,7 +13,7 @@ class UpdateTaskTitleImpl @Inject constructor(
     private val glanceInteractor: GlanceInteractor
 ) : UpdateTaskTitle {
 
-    override suspend fun invoke(taskId: Int?, title: String) {
+    override suspend fun invoke(taskId: Long?, title: String) {
         val task = loadTask(taskId) ?: return
         val updatedTask = task.copy(title = title)
         updateTask(updatedTask)
